+++ packages/typespec-go/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugs Fixed
 
 * Fixed handling of `multipart/form` operations.
+* Fake servers will correctly dispatch calls to all child servers, not just immediate ones.
 
 ## 0.11.0 (2026-04-24)